Move PeekReader out of public interface
Do the wrapping ourselves in DecompressReader.  This means that
DecompressReader.into_inner/get_mut can't return a concrete type, since
it's not possible to unwrap a PeekReader.  Make get_mut() private (we
don't want callers modifying an active stream) and provide into_ methods
for wrapped types with and without Seek.